requirements.txt
requests>=2.28.0
pandas>=1.5.0
PyYAML>=6.0
tenacity>=8.0.1
dbt-core>=1.6.0
dbt-postgres>=1.6.0
pytest>=7.0
python-dateutil>=2.8
